MaxGamer/backend/requirements.txt
# MaxGamer 平台后端依赖
# Flask 核心框架
Flask==3.0.0
Flask-Cors==4.0.0
Flask-SocketIO==5.3.5
python-socketio==5.10.0

# 数据库
SQLAlchemy==2.0.23
Flask-SQLAlchemy==3.1.1

# 认证
Flask-JWT-Extended==4.6.0
bcrypt==4.1.2

# 工具库
python-dotenv==1.0.0
requests==2.31.0

# 短信服务（阿里云）
alibabacloud-dysmsapi20170525>=2.0.24
alibabacloud-tea-openapi>=0.3.9
alibabacloud-tea-util>=0.3.11

# 邮件服务
sendgrid>=6.11.0

# 网络代理
PySocks>=1.7.1

# RAG 知识库系统（AI对话功能）
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy>=1.24.0